Craeted endpoint to load the csv data

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, Float
 from .database import Base
 
 class CO2EmissionRating(Base):
@@ -8,14 +8,14 @@ class CO2EmissionRating(Base):
     brand = Column(String, index=True) #make
     model = Column(String, index=True)
     class_type = Column(String, index=True) # vehicleclass
-    engine_size = Column(Integer, index=True) # enginesize
+    engine_size = Column(Float, index=True) # enginesize
     cylinders = Column(Integer, index=True)
     transmission = Column(String, index=True)
     fueltype = Column(String, index=True)
-    cityconsumption = Column(Integer, index=True)
-    highwayconsumption = Column(Integer, index=True) # hwyconsumption
+    cityconsumption = Column(Float, index=True)
+    highwayconsumption = Column(Float, index=True) # hwyconsumption
     combination_consumption = Column(Integer, index=True) # combconsumption
-    combmpg = Column(Integer, index=True)
+    combmpg = Column(Float, index=True)
     co2emissionrating = Column(String, index=True) # co2emission
